pyproject.toml
[tool.poetry]
name = "stanlib"
version = "0.1.0"
description = "stanlib, a modern python bot focused on provided hackathon information based on your geolocation."
authors = ["Muhammad Bilal Khan <lofi.audit@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymongo = {extras = ["srv"], version = "^4.8.0"}
curl-cffi = "^0.7.1"
rich = "^13.7.1"
python-dotenv = "^1.0.1"
pinecone-client = {extras = ["grpc"], version = "^5.0.1"}
groq = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
